backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pandas>=2.2.3
python-dotenv==1.0.1
python-dateutil==2.9.0
psycopg2-binary>=2.9.11
sqlalchemy==2.0.36
pymssql>=2.3.2
numpy>=1.26.4
